+++ src/failed_attempts_graveyard/Tint.js
