+++ src/helper.js
@@ -2,11 +2,11 @@ export const rgbTo01 = (r, g, b) => {
   return [r / 255, g / 255, b / 255];
 };
 
-function randomIntFromInterval(min, max) {
+export function randomIntFromInterval(min, max) {
   // min and max included
   return Math.floor(Math.random() * (max - min + 1) + min);
 }
-function getRandomFloat(min, max, decimals) {
+export function getRandomFloat(min, max, decimals) {
   const str = (Math.random() * (max - min) + min).toFixed(decimals);
   return parseFloat(str);
 }
