+++ src/materials/ImageFadeMaterial.js
