+++ src/things/Cactus.js
@@ -1,44 +1,28 @@
 import { useRef, useMemo } from 'react';
-import { useFrame, useLoader } from '@react-three/fiber';
-import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader';
+import { useFrame } from '@react-three/fiber';
+import { useGLTF } from '@react-three/drei';
+import { getRandomFloat, randomIntFromInterval } from '../helper';
 
-const MODEL = '/models/good/seaweed.obj';
-function randomIntFromInterval(min, max) {
-  // min and max included
-  return Math.floor(Math.random() * (max - min + 1) + min);
-}
-function getRandomFloat(min, max, decimals) {
-  const str = (Math.random() * (max - min) + min).toFixed(decimals);
-  return parseFloat(str);
-}
+const MODEL = '/models/cactus.glb';
+
 // add more params
-
-/**
- * 
- * idea:
- * add a custom center and x,y range
- * so we can make things like tall in the middle and small on the outside
- * by calculating the distance from the center
- * 
- */
-export function Seaweed({ amount = 30 }) {
-  const geom = useMemo(() => useLoader(OBJLoader, MODEL).children[0].geometry, [])
-  
+export function RandomCacti({ amount = 15 }) {
+  const { nodes, materials } = useGLTF(MODEL);
   const data = useMemo(() => {
     return new Array(amount).fill().map((_, i) => ({
-      x: randomIntFromInterval(-128, 0),
+      x: randomIntFromInterval(-128, 128),
       y: 0,
-      z: randomIntFromInterval(-128, 0),
-      scale: getRandomFloat(1.4, 2),
+      z: randomIntFromInterval(-128, 128),
+      scale: randomIntFromInterval(4, 6),
       rotation: [0, Math.PI / getRandomFloat(Math.PI / 1, 2 * Math.PI), 0],
-      geometry: geom,
+      geometry: nodes.Cactus.geometry,
       elevationOffsetMultiplier: randomIntFromInterval(2, 5),
       timeFrequency: randomIntFromInterval(22, 44) / 10000,
     }));
-  }, [geom]);
+  }, []);
 
   return data.map((props, i) => (
-    <WobbleSeaweed key={i} {...props} position={[props.x, props.y, props.z]} />
+    <WobbleMesh key={i} {...props} position={[props.x, props.y, props.z]} />
   ));
 }
 
@@ -51,14 +35,8 @@ export function Seaweed({ amount = 30 }) {
  * @param {Vector3} position [0,0,0]
  * @returns
  */
-export function WobbleSeaweed(props) {
-  const {
-    scale,
-    geometry,
-    elevationOffsetMultiplier,
-    timeFrequency,
-    rotation,
-  } = props;
+export function WobbleMesh(props) {
+  const { scale, geometry, elevationOffsetMultiplier, timeFrequency, rotation } = props;
   const uTime = useRef({ value: 0.0 });
   // Update cactus time uniform
   useFrame(
@@ -66,13 +44,13 @@ export function WobbleSeaweed(props) {
       (uTime.current.value = clock.elapsedTime * 1000 + timeFrequency)
   );
   return (
-    <group {...props} dispose={null}>
+    <group {...props}  dispose={null}>
       <mesh
         castShadow
         geometry={geometry}
+        scale={scale}
         position={[0, 0, 0]}
-        rotation={[0, 0, 0]}
-        scale={1}
+        rotation={[Math.PI / 2, 0, 0]}
       >
         <meshStandardMaterial
           attach="material"
@@ -112,8 +90,6 @@ export function WobbleSeaweed(props) {
             );
           }}
           color="green"
-          opacity={0.9}
-          transparent={true}
           roughness={1.0}
           metalness={0.0}
         />
@@ -121,3 +97,6 @@ export function WobbleSeaweed(props) {
     </group>
   );
 }
+
+export default WobbleMesh;
+useGLTF.preload(MODEL);
