+++ src/things/Rock.js
@@ -1,28 +1,23 @@
 import { useRef, useMemo } from 'react';
 import { useFrame, useLoader } from '@react-three/fiber';
 import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader';
+import { getRandomFloat, randomIntFromInterval } from '../helper';
 
-const MODEL = '/models/good/Monument.obj';
-function randomIntFromInterval(min, max) {
-  // min and max included
-  return Math.floor(Math.random() * (max - min + 1) + min);
-}
+const MODEL = '/models/good/rock.obj';
 
-function getRandomFloat(min, max, decimals=2) {
-  const str = (Math.random() * (max - min) + min).toFixed(decimals);
-  return parseFloat(str);
-}
 // add more params
-export function Monument({ amount = 25 }) {
-  const geom = useMemo(() => useLoader(OBJLoader, MODEL).children[0].geometry, [])
-
+export function Rocks({ amount = 25 }) {
+  const geom = useMemo(
+    () => useLoader(OBJLoader, MODEL).children[0].geometry,
+    []
+  );
   const data = useMemo(() => {
     return new Array(amount).fill().map((_, i) => ({
       x: randomIntFromInterval(-128, 128),
-      y: -10,
+      y: 0-3,
       z: randomIntFromInterval(-128, 128),
-      scale: 2,
-      rotation: [0, Math.PI / getRandomFloat(Math.PI / 1, 2 * Math.PI), 0],
+      scale: getRandomFloat(2,3),
+      rotation: [0,  getRandomFloat(Math.PI / 1, 2 * Math.PI), 0],
       geometry: geom,
       elevationOffsetMultiplier: randomIntFromInterval(2, 5),
       timeFrequency: randomIntFromInterval(22, 44) / 10000,
@@ -30,7 +25,7 @@ export function Monument({ amount = 25 }) {
   }, [geom]);
 
   return data.map((props, i) => (
-    <AMonument key={i} {...props} position={[props.x, props.y, props.z]} />
+    <ARock key={i} {...props} position={[props.x, props.y, props.z]} />
   ));
 }
 
@@ -43,26 +38,20 @@ export function Monument({ amount = 25 }) {
  * @param {Vector3} position [0,0,0]
  * @returns
  */
-export function AMonument(props) {
+export function ARock(props) {
   const { scale, geometry, rotation } = props;
-  const rgbTo01=(r,g,b)=> {
-    return [r/255,g/255,b/255];
-  }
   return (
     <group {...props} dispose={null}>
       <mesh
         castShadow
         geometry={geometry}
         scale={scale}
-        position={[0, 10 / scale, 0]}
+        position={[0, 0, 0]}
       >
         <meshStandardMaterial
-          attach="material"
-          toneMapped={false}
-          emissive={true}
-          color={rgbTo01(255*1,215,0)}
-          roughness={0.3}
-          metalness={1.0}
+          color="darkgrey"
+          roughness={0.9}
+          metalness={0.1}
         />
       </mesh>
     </group>
