+++ src/things/Seaweed.js
@@ -1,34 +1,36 @@
 import { useRef, useMemo } from 'react';
-import { useFrame } from '@react-three/fiber';
-import { useGLTF } from '@react-three/drei';
+import { useFrame, useLoader } from '@react-three/fiber';
+import { OBJLoader } from 'three/examples/jsm/loaders/OBJLoader';
+import { getRandomFloat, randomIntFromInterval } from '../helper';
 
-const MODEL = '/models/cactus.glb';
-function randomIntFromInterval(min, max) {
-  // min and max included
-  return Math.floor(Math.random() * (max - min + 1) + min);
-}
-function getRandomFloat(min, max, decimals) {
-  const str = (Math.random() * (max - min) + min).toFixed(decimals);
-  return parseFloat(str);
-}
-// add more params
-export function RandomCacti({ amount = 15 }) {
-  const { nodes, materials } = useGLTF(MODEL);
+const MODEL = '/models/good/seaweed.obj';
+
+/**
+ * 
+ * idea:
+ * add a custom center and x,y range
+ * so we can make things like tall in the middle and small on the outside
+ * by calculating the distance from the center
+ * 
+ */
+export function Seaweed({ amount = 30 }) {
+  const geom = useMemo(() => useLoader(OBJLoader, MODEL).children[0].geometry, [])
+  
   const data = useMemo(() => {
     return new Array(amount).fill().map((_, i) => ({
-      x: randomIntFromInterval(-128, 128),
+      x: randomIntFromInterval(-128, 0),
       y: 0,
-      z: randomIntFromInterval(-128, 128),
-      scale: randomIntFromInterval(4, 6),
+      z: randomIntFromInterval(-128, 0),
+      scale: getRandomFloat(1.4, 2),
       rotation: [0, Math.PI / getRandomFloat(Math.PI / 1, 2 * Math.PI), 0],
-      geometry: nodes.Cactus.geometry,
+      geometry: geom,
       elevationOffsetMultiplier: randomIntFromInterval(2, 5),
       timeFrequency: randomIntFromInterval(22, 44) / 10000,
     }));
-  }, []);
+  }, [geom]);
 
   return data.map((props, i) => (
-    <WobbleMesh key={i} {...props} position={[props.x, props.y, props.z]} />
+    <WobbleSeaweed key={i} {...props} position={[props.x, props.y, props.z]} />
   ));
 }
 
@@ -41,8 +43,14 @@ export function RandomCacti({ amount = 15 }) {
  * @param {Vector3} position [0,0,0]
  * @returns
  */
-export function WobbleMesh(props) {
-  const { scale, geometry, elevationOffsetMultiplier, timeFrequency, rotation } = props;
+export function WobbleSeaweed(props) {
+  const {
+    scale,
+    geometry,
+    elevationOffsetMultiplier,
+    timeFrequency,
+    rotation,
+  } = props;
   const uTime = useRef({ value: 0.0 });
   // Update cactus time uniform
   useFrame(
@@ -50,13 +58,13 @@ export function WobbleMesh(props) {
       (uTime.current.value = clock.elapsedTime * 1000 + timeFrequency)
   );
   return (
-    <group {...props}  dispose={null}>
+    <group {...props} dispose={null}>
       <mesh
         castShadow
         geometry={geometry}
-        scale={scale}
         position={[0, 0, 0]}
-        rotation={[Math.PI / 2, 0, 0]}
+        rotation={[0, 0, 0]}
+        scale={1}
       >
         <meshStandardMaterial
           attach="material"
@@ -96,6 +104,8 @@ export function WobbleMesh(props) {
             );
           }}
           color="green"
+          opacity={0.9}
+          transparent={true}
           roughness={1.0}
           metalness={0.0}
         />
@@ -103,6 +113,3 @@ export function WobbleMesh(props) {
     </group>
   );
 }
-
-export default WobbleMesh;
-useGLTF.preload(MODEL);
