+++ src/things/fish-swarm/SimpleSlider.js
